Revert "MMT-1213 remove blank fields in metadata from bulk update submissions"
This reverts commit c555556a8b936c7ababac3ff69d5434fe00ff39b.
